# services/intelligence/contracts/phase2/openapi.yaml
openapi: 3.1.0
info:
  title: Study Tracker Phase 2 Intelligence API
  version: 1.1.0
  description: Public, owner-scoped contract. Hidden answer keys, rubrics, reference solutions, and hidden tests are server-only.
  x-contract-status: approved
servers:
  - url: /
paths:
...
  /v1/materials/{materialId}/ingestion:
    get:
      operationId: getIngestionStatus
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RequestId' }, { $ref: '#/components/parameters/MaterialId' }]
      responses: { '200': { description: Ingestion status, content: { application/json: { schema: { $ref: '#/components/schemas/IngestionStatus' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' } }
  /v1/materials/{materialId}/content:
    get:
      operationId: getMaterialContentPreview
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/RequestId' }, { $ref: '#/components/parameters/MaterialId' }]
      responses: { '200': { description: Partial extracted content preview, content: { application/json: { schema: { $ref: '#/components/schemas/MaterialContentPreview' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase access token; sub is the ownership key. }
  parameters:
    RequestId: { in: header, name: X-Request-ID, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
    MaterialId: { in: path, name: materialId, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
  responses:
...
    Unauthorized: { description: Missing or invalid JWT, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceError' } } } }
...
    NotFound: { description: Object not found in owner scope, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceError' } } } }
...
  schemas:
    Id: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$' }
    Timestamp: { type: string, format: date-time }
...
    IngestionStatus: { type: object, additionalProperties: false, required: [materialId, state, progress, updatedAt], properties: { materialId: { $ref: '#/components/schemas/Id' }, state: { type: string }, progress: { type: number, minimum: 0, maximum: 1 }, attempt: { type: integer, minimum: 1 }, retryable: { type: boolean }, error: { $ref: '#/components/schemas/ServiceError' }, updatedAt: { $ref: '#/components/schemas/Timestamp' } } }
    MaterialContentPreview: { type: object, additionalProperties: false, required: [materialId, state, updatedAt], properties: { materialId: { $ref: '#/components/schemas/Id' }, state: { type: string }, previewText: { type: string, maxLength: 10000 }, chunkCount: { type: integer, minimum: 0 }, ready: { type: boolean }, updatedAt: { $ref: '#/components/schemas/Timestamp' } } }
...
    ServiceError: { type: object, additionalProperties: false, required: [code, message, requestId, retryable], properties: { code: { type: string, enum: [invalid_request, unauthorized, forbidden, not_found, conflict, safety_block, quota_exhausted, provider_unavailable, provider_timeout, malformed_output, validation_failed, internal_error] }, message: { type: string }, requestId: { $ref: '#/components/schemas/Id' }, correlationId: { $ref: '#/components/schemas/Id' }, retryable: { type: boolean }, retryAfterSeconds: { type: integer, minimum: 1 } } }